feat: remove old UserPage, related older actions and version toggle
