+++ src/index.jsx
@@ -13,9 +13,8 @@ import { Switch, Route } from 'react-router-dom';
 import SupportHomePage from './supportHome/SupportHomePage';
 import Header from './supportHeader';
 import appMessages from './i18n';
-import UserPage from './users/UserPage';
 import SupportToolsTab from './SupportToolsTab/SupportToolsTab';
-import UserPageV2 from './users/v2/UserPage';
+import UserPageV2 from './users/UserPage';
 import FBEIndexPage from './FeatureBasedEnrollments/FeatureBasedEnrollmentIndexPage';
 import UserMessagesProvider from './userMessages/UserMessagesProvider';
 import ProgramEnrollmentsIndexPage from './ProgramEnrollments/ProgramEnrollmentsIndexPage';
@@ -38,7 +37,6 @@ subscribe(APP_READY, () => {
         <Header />
         <Switch>
           <Route exact path="/" component={SupportHomePage} />
-          <Route path="/users" component={UserPage} />
           <Route path="/v2" component={SupportToolsTab} />
           <Route path="/v2/learner_information" component={UserPageV2} />
           <Route path="/v2/feature_based_enrollments" component={FBEIndexPage} />
