+++ src/users/UserPage.jsx
@@ -1,22 +1,27 @@
-import { history } from '@edx/frontend-platform';
+import { camelCaseObject, history } from '@edx/frontend-platform';
 import PropTypes from 'prop-types';
 import React, {
-  useCallback, useContext, useEffect, useState,
+  useCallback, useContext, useEffect, useState, useLayoutEffect,
 } from 'react';
-import { Link } from 'react-router-dom';
 import PageLoading from '../components/common/PageLoading';
 import AlertList from '../userMessages/AlertList';
 import { USER_IDENTIFIER_INVALID_ERROR } from '../userMessages/messages';
 import UserMessagesContext from '../userMessages/UserMessagesContext';
-import {
-  isEmail, isValidUsername, isValidLMSUserID, extractParams,
-} from '../utils/index';
+import { isEmail, isValidUsername, isValidLMSUserID } from '../utils/index';
 import { getAllUserData } from './data/api';
 import UserSearch from './UserSearch';
+import LearnerInformation from './v2/LearnerInformation';
+import { LEARNER_INFO_TAB, TAB_PATH_MAP } from '../SupportToolsTab/constants';
 
 // Supports urls such as /users/?username={username}, /users/?email={email} and /users/?lms_user_id={lms_user_id}
 export default function UserPage({ location }) {
-  const params = extractParams(location.search);
+  // converts query params from url into map e.g. ?param1=value1&param2=value2 -> {param1: value1, param2=value2}
+  const params = new Map(
+    location.search
+      .slice(1) // removes '?' mark from start
+      .split('&')
+      .map(queryParams => queryParams.split('=')),
+  );
 
   if (params.has('email')) {
     const email = params.get('email');
@@ -27,6 +32,7 @@ export default function UserPage({ location }) {
     params.get('username') || params.get('email') || params.get('lms_user_id') || undefined,
   );
   const [searching, setSearching] = useState(false);
+  const [data, setData] = useState({ enrollments: null, entitlements: null });
   const [loading, setLoading] = useState(false);
   const { add, clear } = useContext(UserMessagesContext);
 
@@ -37,7 +43,7 @@ export default function UserPage({ location }) {
   }
 
   function getUpdatedURL(value) {
-    const updatedHistory = `/users/?PARAM_NAME=${value}`;
+    const updatedHistory = `${TAB_PATH_MAP['learner-information']}/?PARAM_NAME=${value}`;
     let identifierType = '';
 
     if (isEmail(value)) {
@@ -54,7 +60,7 @@ export default function UserPage({ location }) {
   function processSearchResult(searchValue, result) {
     if (result.errors.length > 0) {
       result.errors.forEach((error) => add(error));
-      history.replace('/users');
+      history.replace(`${TAB_PATH_MAP['learner-information']}`);
       document.title = 'Support Tools | edX';
     } else {
       pushHistoryIfChanged(getUpdatedURL(searchValue));
@@ -66,7 +72,7 @@ export default function UserPage({ location }) {
   }
 
   function validateInput(input) {
-    if (!isValidUsername(input) && !isEmail(input)) {
+    if (!isValidUsername(input) && !isEmail(input) && !isValidLMSUserID(input)) {
       clear('general');
       add({
         code: null,
@@ -75,7 +81,7 @@ export default function UserPage({ location }) {
         type: 'error',
         topic: 'general',
       });
-      history.replace('/users');
+      history.replace(`${TAB_PATH_MAP['learner-information']}`);
       return false;
     }
     return true;
@@ -90,15 +96,20 @@ export default function UserPage({ location }) {
       setUserIdentifier(searchValue);
       setLoading(true);
       getAllUserData(searchValue).then((result) => {
+        setData(camelCaseObject(result));
         processSearchResult(searchValue, result);
       });
       // This is the case of an empty search (maybe a user wanted to clear out what they were seeing)
     } else if (searchValue === '') {
       clear('general');
-      history.replace('/users');
+      history.replace(`${TAB_PATH_MAP['learner-information']}`);
       setLoading(false);
       setSearching(false);
     }
+  });
+  const handleUserSummaryChange = useCallback(() => {
+    setSearching(true);
+    handleFetchSearchResults(userIdentifier);
   });
 
   const handleSearchInputChange = useCallback((searchValue) => {
@@ -122,11 +133,17 @@ export default function UserPage({ location }) {
     }
   }, [params.get('username'), params.get('email'), params.get('lms_user_id')]);
 
+  // To change the url with appropriate query param if query param info is not present in URL
+  useLayoutEffect(() => {
+    if (userIdentifier
+      && location.pathname.indexOf(TAB_PATH_MAP[LEARNER_INFO_TAB]) !== -1
+      && !(params.get('email') || params.get('username') || params.get('lms_user_id'))) {
+      pushHistoryIfChanged(getUpdatedURL(userIdentifier));
+    }
+  });
+
   return (
-    <main className="container-fluid mt-3 mb-5">
-      <section className="mb-3">
-        <Link to="/">&lt; Back to Tools</Link>
-      </section>
+    <main className="mt-3 mb-5">
       <AlertList topic="general" className="mb-3" />
       {/* NOTE: the "key" here causes the UserSearch component to re-render completely when the
       user identifier changes.  Doing so clears out the search box. */}
@@ -136,10 +153,11 @@ export default function UserPage({ location }) {
         searchHandler={handleSearchInputChange}
       />
       {loading && <PageLoading srMessage="Loading" />}
-      {!loading && !userIdentifier && (
-        <section>
-          <p>Please search for a username or email.</p>
-        </section>
+      {!loading && data.user && data.user.username && (
+        <LearnerInformation
+          user={data.user}
+          changeHandler={handleUserSummaryChange}
+        />
       )}
     </main>
   );
