--- src/users/data/test/idvStatus.js
@@ -1,8 +1,0 @@
-const idvStatusData = {
-  status: 'verified',
-  expirationDatetime: null,
-  isVerified: true,
-  extraData: [],
-};
-
-export default idvStatusData;
