+++ src/users/v2/OnboardingStatus.test.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { waitForComponentToPaint } from '../../setupTest';
 import OnboardingStatus from './OnboardingStatus';
 import UserMessagesProvider from '../../userMessages/UserMessagesProvider';
-import { v2OnboardingStatusData as onboardingStatusData } from '../data/test/onboardingStatus';
+import v2OnboardingStatusData from '../data/test/onboardingStatus';
 import enrollmentsData from '../data/test/enrollments';
 import { titleCase, formatDate } from '../../utils';
 
@@ -22,7 +22,7 @@ describe('Onboarding Status', () => {
   };
 
   beforeEach(async () => {
-    jest.spyOn(api, 'getV2OnboardingStatus').mockImplementationOnce(() => Promise.resolve(onboardingStatusData));
+    jest.spyOn(api, 'getV2OnboardingStatus').mockImplementationOnce(() => Promise.resolve(v2OnboardingStatusData));
     wrapper = mount(<OnboardingStatusWrapper {...props} />);
     await waitForComponentToPaint(wrapper);
   });
@@ -37,17 +37,19 @@ describe('Onboarding Status', () => {
     const verifiedInData = wrapper.find('Table#verified-in-data');
     const verifiedInDataBody = verifiedInData.find('tbody tr td');
     expect(verifiedInDataBody).toHaveLength(4);
-    expect(verifiedInDataBody.at(0).text()).toEqual(onboardingStatusData.verifiedIn.courseId);
-    expect(verifiedInDataBody.at(1).text()).toEqual(titleCase(onboardingStatusData.verifiedIn.onboardingStatus));
-    expect(verifiedInDataBody.at(2).text()).toEqual(formatDate(onboardingStatusData.verifiedIn.expirationDate));
+    expect(verifiedInDataBody.at(0).text()).toEqual(v2OnboardingStatusData.verifiedIn.courseId);
+    expect(verifiedInDataBody.at(1).text()).toEqual(titleCase(v2OnboardingStatusData.verifiedIn.onboardingStatus));
+    expect(verifiedInDataBody.at(2).text()).toEqual(formatDate(v2OnboardingStatusData.verifiedIn.expirationDate));
     expect(verifiedInDataBody.at(3).text()).toEqual('Link');
 
     const currentStatusData = wrapper.find('Table#current-status-data');
     const currentStatusDataBody = currentStatusData.find('tbody tr td');
     expect(currentStatusDataBody).toHaveLength(4);
-    expect(currentStatusDataBody.at(0).text()).toEqual(onboardingStatusData.currentStatus.courseId);
-    expect(currentStatusDataBody.at(1).text()).toEqual(titleCase(onboardingStatusData.currentStatus.onboardingStatus));
-    expect(currentStatusDataBody.at(2).text()).toEqual(formatDate(onboardingStatusData.currentStatus.expirationDate));
+    expect(currentStatusDataBody.at(0).text()).toEqual(v2OnboardingStatusData.currentStatus.courseId);
+    expect(currentStatusDataBody.at(1).text()).toEqual(
+      titleCase(v2OnboardingStatusData.currentStatus.onboardingStatus),
+    );
+    expect(currentStatusDataBody.at(2).text()).toEqual(formatDate(v2OnboardingStatusData.currentStatus.expirationDate));
     expect(currentStatusDataBody.at(3).text()).toEqual('Link');
   });
 
@@ -78,7 +80,7 @@ describe('Onboarding Status', () => {
   });
 
   it('No Onboarding Status Data with current status data only', async () => {
-    const onboardingData = { ...onboardingStatusData, verifiedIn: null };
+    const onboardingData = { ...v2OnboardingStatusData, verifiedIn: null };
 
     jest.spyOn(api, 'getEnrollments').mockImplementationOnce(() => Promise.resolve(enrollmentsData));
     jest.spyOn(api, 'getV2OnboardingStatus').mockImplementationOnce(() => Promise.resolve(onboardingData));
@@ -91,9 +93,11 @@ describe('Onboarding Status', () => {
     const currentStatusData = wrapper.find('Table#current-status-data');
     const currentStatusDataBody = currentStatusData.find('tbody tr td');
     expect(currentStatusDataBody).toHaveLength(4);
-    expect(currentStatusDataBody.at(0).text()).toEqual(onboardingStatusData.currentStatus.courseId);
-    expect(currentStatusDataBody.at(1).text()).toEqual(titleCase(onboardingStatusData.currentStatus.onboardingStatus));
-    expect(currentStatusDataBody.at(2).text()).toEqual(formatDate(onboardingStatusData.currentStatus.expirationDate));
+    expect(currentStatusDataBody.at(0).text()).toEqual(v2OnboardingStatusData.currentStatus.courseId);
+    expect(currentStatusDataBody.at(1).text()).toEqual(
+      titleCase(v2OnboardingStatusData.currentStatus.onboardingStatus),
+    );
+    expect(currentStatusDataBody.at(2).text()).toEqual(formatDate(v2OnboardingStatusData.currentStatus.expirationDate));
     expect(currentStatusDataBody.at(3).text()).toEqual('Link');
   });
 
@@ -109,7 +113,7 @@ describe('Onboarding Status', () => {
       enrollmentDate: null,
       instructorDashboardLink: null,
     };
-    const onboardingData = { ...onboardingStatusData, verifiedIn: nullData };
+    const onboardingData = { ...v2OnboardingStatusData, verifiedIn: nullData };
 
     jest.spyOn(api, 'getV2OnboardingStatus').mockImplementationOnce(() => Promise.resolve(onboardingData));
     wrapper = mount(<OnboardingStatusWrapper {...props} />);
